+++ example/lib/pages/select_form_field.dart
@@ -95,7 +95,7 @@ class _SelectFormFieldPageState extends State<SelectFormFieldPage> {
                 ),
                 const SizedBox(height: 16),
                 ShadButton(
-                  text: const Text('Submit'),
+                  child: const Text('Submit'),
                   onPressed: () {
                     print('submitted');
                     if (formKey.currentState!.saveAndValidate()) {