+++ example/lib/pages/tabs.dart
@@ -20,12 +20,11 @@ class TabsPage extends StatelessWidget {
           tabs: [
             ShadTab(
               value: 'account',
-              text: const Text('Account'),
               content: ShadCard(
                 title: const Text('Account'),
                 description: const Text(
                     "Make changes to your account here. Click save when you're done."),
-                footer: const ShadButton(text: Text('Save changes')),
+                footer: const ShadButton(child: Text('Save changes')),
                 child: Column(
                   mainAxisSize: MainAxisSize.min,
                   children: [
@@ -43,15 +42,15 @@ class TabsPage extends StatelessWidget {
                   ],
                 ),
               ),
+              child: const Text('Account'),
             ),
             ShadTab(
               value: 'password',
-              text: const Text('Password'),
               content: ShadCard(
                 title: const Text('Password'),
                 description: const Text(
                     "Change your password here. After saving, you'll be logged out."),
-                footer: const ShadButton(text: Text('Save password')),
+                footer: const ShadButton(child: Text('Save password')),
                 child: Column(
                   children: [
                     const SizedBox(height: 16),
@@ -68,6 +67,7 @@ class TabsPage extends StatelessWidget {
                   ],
                 ),
               ),
+              child: const Text('Password'),
             ),
           ],
         ),