+++ lib/src/components/tabs.dart
@@ -361,7 +361,7 @@ class ShadTab<T> extends StatefulWidget implements PreferredSizeWidget {
   const ShadTab({
     super.key,
     required this.value,
-    required this.text,
+    required this.child,
     this.content,
     this.icon,
     this.enabled = true,
@@ -416,7 +416,7 @@ class ShadTab<T> extends StatefulWidget implements PreferredSizeWidget {
   final T value;
 
   /// The text of the tab.
-  final Widget text;
+  final Widget child;
 
   /// The content of the tab.
   final Widget? content;
@@ -731,10 +731,6 @@ class _ShadTabState<T> extends State<ShadTab<T>> {
           foregroundColor: selected
               ? effectiveSelectedForegroundColor
               : effectiveForegroundColor,
-          text: DefaultTextStyle(
-            style: theme.textTheme.small,
-            child: widget.text,
-          ),
           shadows: selected ? effectiveSelectedShadows : effectiveShadows,
           onPressed: () {
             inherited.controller.select(widget.value);
@@ -769,6 +765,10 @@ class _ShadTabState<T> extends State<ShadTab<T>> {
           onDoubleTapDown: widget.onDoubleTapDown,
           onDoubleTapCancel: widget.onDoubleTapCancel,
           longPressDuration: widget.longPressDuration,
+          child: DefaultTextStyle(
+            style: theme.textTheme.small,
+            child: widget.child,
+          ),
         );
       },
     );